remove circular deps
