+++ app/src/__tests__/registry.test.ts
@@ -1,3 +1,6 @@
+import { Ethereum } from '@/registry/mainnet/chains'
+import { routes } from '@/registry/mainnet/routes'
+import { EthereumTokens } from '@/registry/mainnet/tokens'
 import { error, log } from 'console'
 
 error('This is an error')
@@ -5,7 +8,8 @@ log('Something to log')
 
 describe('Token Transfer Routes Registry', () => {
   test('routes should be defined', async () => {
-    const { EthereumTokens } = await import('../registry/mainnet/tokens') // ✅ Fixes circular issue
     console.log('Eth:', EthereumTokens.ETH)
+    console.log('route:', routes.at(1))
+    console.log('chain:', Ethereum)
   })
 })
